@@ -8,7 +8,7 @@ namespace VpnHood.Common.Converters
     public class IPAddressConverter : JsonConverter<IPAddress>
     {
         public override IPAddress Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            => IPAddress.Parse(reader.GetString());
+            => IPAddress.Parse(reader.GetString() ?? "");
         public override void Write(Utf8JsonWriter writer, IPAddress value, JsonSerializerOptions options)
             => writer.WriteStringValue(value.ToString());
     }
