@@ -9,14 +9,14 @@ namespace VpnHood.Common.Converters
     public class IPEndPointConverter : JsonConverter<IPEndPoint>
     {
         public override IPEndPoint Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            => Parse(reader.GetString() ?? throw new InvalidOperationException("Could not deserialize an IPEndPoint!"));
+            => Parse(reader.GetString() ?? "");
         public override void Write(Utf8JsonWriter writer, IPEndPoint value, JsonSerializerOptions options)
             => writer.WriteStringValue(value.ToString());
 
         public static IPEndPoint Parse(string value)
         {
             if (!TryParse(value, out var ipEndPoint))
-                throw new ArgumentException($"Could not parse {value} to an {nameof(IPEndPoint)}!");
+                throw new ArgumentException($"Could not parse {nameof(IPEndPoint)} from: {value}!");
             return ipEndPoint;
         }
 
