@@ -7,7 +7,7 @@ namespace VpnHood.Common.Converters
     public class VersionConveter : JsonConverter<Version>
     {
         public override Version Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            => Version.Parse(reader.GetString());
+            => Version.Parse(reader.GetString() ?? "");
         public override void Write(Utf8JsonWriter writer, Version value, JsonSerializerOptions options)
             => writer.WriteStringValue(value.ToString());
     }
